Add guidance on how to find valid xrefs to reference.
It turns out that the syntax for xrefs is super hard and not actually easy to find. However, running docfx locally will generate a convenient map, which takes away the memorization process (you can just Ctrl+F for thing you want to link to)

Adding this nugget of guidance to our dev docs.