Input recording editor window (#5150)
Stage 2 of the input recording feature: Editor tools for recording and playing back input animation.

This provides editor-level GUI for the recording and playback services.

InputRecordingWindow is a standalone window that can be opened from the Mixed Reality Toolkit menu.

It has two modes:
* Record shows a recording button that toggles the recording service. When recording is stopped, a popup is shown to select a file destination for the recorded data.
* Playback allows loading recorded input from a file. The data is then applied back through the input playback service.

The Input Recording Window is an alternative to the toggle button in the HandTracking examples. Both of these will work independently by toggling the input recording service.