Update the spatial awareness architecture doc.
The old spatial awareness architecture doc got fairly specific into the specific interfaces and classes and events, which is really helpful if you already have an understanding of the system (i.e. it provides a reference to things you already know).

However, if you don't know about how everything works, what you really want is a high level introduction to what it is, how it's roughly structured, and what each rough piece is responsible for.

I also called out the specific platforms that we have inbox support for.