Logging of EyeGaze data to csv file is culture-specific
https://github.com/microsoft/MixedRealityToolkit-Unity/issues/5184

On certain locales, the CSV delimiter isn't the comma, but the semicolon instead.

Previously, we assumed that the delimiter would always be the comma, but this would cause problems in cases like de-DE where numbers are written with commas (instead of decimals) to represent the decimal digits.

So for example, en_US:

123.123, ABCABC, XYZXYZ

Without this change, it would look like this on de_DE
123,123, ABCABC, XYZXYZ

This is obviously totally wrong - now we have four cells, it doesn't make sense!

The solution is to use the real delimiter for the current locale:
123,123; ABCABC; XYZXYZ