Validate collider type in NearInteractionGrabbable (#5211)
Grab behavior implementation in FocusProvider relies on Unity Collider.GetClosestPoint to select a closest object, when there are several within a grab distance. From docs, https://docs.unity3d.com/ScriptReference/Physics.ClosestPoint.html , this call requires collider to be of one of 4 types. If collider is not, GetClosestPoint would emit a warning and return 0, which will break grab logic. As this behavior is outside of Mrtk scope, adding explicit exception on initialization of Grabbable component should at least make this error visible.