Fix Unity UI touchable "greedy" activation.
Physics-based touchables now have a broadphase that limits the number of
touchables that need to be examined for finding the closest one.
Unity UI touchables don't have such a broadphase test, so distances needs
to be checked against the touch distance as well, otherwise any UnityUI
touchable will activate if no other touchable is closer.