Add export and import of input animation data to and from glTF files.
This supplements (and should ultimately replace) the custom binary format, and provides a standardized way to exchange
recorded input data with other tools.

A new set of utility classes has been added to serialize InputAnimation curves into a binary buffer suitable for glTF
accessors. The existing GltfObject in MRTK is constructed with simple nodes that represent the camera and hand
controllers. A matching importer finds matching nodes in any glTF file to try and interpret the animation data as MRTK
input again.

The Unity JsonUtility class is not flexible enough to handle details of the glTF schema, such as minimum array length,
or whether enums should be exported by name or by int value. For this reason an alternative JSON builder and parser pair
has been added. These can handle JsonAttributes, which specify schema details as part of the C# reflection, so that the
GltfObject can still be exported as "structured JSON".