Add a link.xml to Assets to prevent critical MRTK components from getting optimized away in IL2CPP builds
https://github.com/microsoft/MixedRealityToolkit-Unity/issues/4265

See this comment in that issue for some more details on what happened here:
https://github.com/microsoft/MixedRealityToolkit-Unity/issues/4265#issuecomment-510246866

Also, in case you were wondering how this was generated, these are the assemblies that are referenced in the default MRTK profiles (i.e. look for things of type "SystemType" and look at their corresponding references). In the long run this has to be more automated, as this was obviously a highly manual process. See this issue (https://github.com/microsoft/MixedRealityToolkit-Unity/issues/5273) for more details on that.