Update the build scripts to produce Foundation and Examples .unitypackages.
build.ps1 used to be used to do both nuget packing and unity package generation. It hasn't been used since we moved NuGet pack to the YAML pipelines themselves, the way it worked before anyways was by looking at the directory that nuspec files were in, which is not valid for packages like the Foundation one (which contains a number of other folders).

This change removes the old build.ps1 script in favor of having a script that is explicitly about building MRTK .unitypackage files. Inside you will find the configuration itself (i.e. the set of packages along with their corresponding folders)