Update the IMixedRealityRaycastProvider to support custom graphics raycast functionality.
https://github.com/microsoft/MixedRealityToolkit-Unity/issues/4917

This change adds an extension point to the IMixedRealityRaycastProvider to allow for customizing how the GraphicsRaycast works. This allows for clients to customize the priority order of raycast results.

The goal here was to make it possible for clients to change the order of RaycastResultComparer::Comparers, so the changes involved:

1) Adding the GraphicsRaycast function to IMixedRealityRaycastProvider.
2) Updating RaycastResultComparer's Comparers comparison functions so that they can be referenced in subclasses.
3) Updating the default implementation to do exactly what it does today.

Note that in theory it would be possible to accomplish this by just doing #1 (this would require callers to do a bunch of duplicated work, so I opted to do #2-#3 as well)