Update the configuration guide to remove references to things coming soon.
The doc contains a lot of references to things that are coming soon, and while we have items in the backlog to improve these, not all of these are going to be coming in the immediate future. It's misleading and confidence-draining to see things always listed as coming soon, so it's better to have our docs just say what the current state of things are.