Update the assert text to be more helpful, and update the doccomments of the focus handler interfaces.
I was spending time updating the Periodic Table app from RC2 -> GA, and while mostly things were working, I noticed that this assert was firing (it was added after RC2.1) pretty much every time you had a near pointer around the molecule object. It turns out, there's an issue in the heirarchy of the app itself that leads to this issue, but I was initially super confused about this, since this assert seems bad, but there's really not that much guidance on how to fix it and what's wrong.

To that end, I've updated the assert to suggest where to look (i.e. the common case) and also updated the doccomments associated with the corresponding interfaces to provide more information about both state transition mechanics, and also the general warning to be careful about having different nesting mechanics and pitfalls