Adds the MixedRealityToolkit.Tools folder and package.
There are a number of features currently in the the MixedRealityToolkit folder that are useful at editor and build time, but are otherwise never going to be shipping with the final application. For example, the screenshot utility is incredibly useful for people who want to take screenshots of their features (useful for MRTK devs especially!) but otherwise is a bit odd to have directly in the MRTK "core" folder, because it's not really a core part of a consumer's development process.

This change spawns a new folder to put these things in (MixedRealityToolkit.Tools) where we can place features (especially ones that are editor-only) that accrue value during the development cycle (or could just generally be useful tools even outside of the MRTK, but due to repo/CI resourcing we're currently putting them in here), but otherwise don't ship with the app itself.

Note that this change doesn't move EVERYTHING over (for example, the BuildWindow tool is still left inside the core part because there is some history there and it would be some degree of a "package" breaking change to move that over.)

This change also adds the new MixedRealityToolkit.Tools.nuget package, along with the corresponding MixedRealityToolkit.Tools.unitypackage