Fixes gesture pinch release mechanics involving two mouse button interactions.
Prior to this fix, when using gestures (GGV) style interactions, you would run into this bug:

1) Press space bar to engage the hand simulation.
2) Hold down left mouse to engage pinch.
3) Hold down the right mouse to rotate the camera.
4) Lift the left mouse to disengage pinch.

Step #4 wouldn't work (it would stay engaged in the pinch until you lifted the right mouse button) because the existing code would see that the right mouse button was down, and even though it was not configured to do anything (i.e. a configuration of "none" means "don't affect the state of the gesture simulation"). However because the function returned "none", this told the rest of the code to keep doing what it was doing before (which was, simulating the pinch gesture associated with the left mouse click.)

You have to return the default (or open) pose explicitly to get the hand to open, and by making the code change to actually honor the lack of configuration, this all works out.