Add a somewhat helpful message on CreateInstance failure.
It turns out that the existing logging isn't very helpful, because the inner message isn't logged. Thus, if the client here is creating something and has an exception in their code, all they will see right now is a TargetInvocationException, instead of a real callstack associated with their code.

This is SUPER hard to debug (you have to know to turn on all managed exceptions) so hopefully this increased logging will help clients figure out what happened when we try to new their objects up.