Resolve the ArgumentException error in MRTK+Unity startup
When ingesting MRTK for the first time (especially via git) you'll see an ArgumentException warning, which happens as MRTK is figuring out where its files are being placed. This shows as an error but is actualy benign (i.e. it happens in some processing of UPM packages that have no effect on the MRTK).

The cause here was an overly aggressive .Replace call - it replaced ALL instances of "Asset" with the application asset path (even when Asset was showing up much later in the path). It was really only intending to replace the prefix "Asset" string, so this fixes it to only do that.