Fixes build error when a scene with MRTK instance is open in editor
https://github.com/microsoft/MixedRealityToolkit-Unity/issues/5131

This code shouldn't be running during player builds - it's intended to only run during edit mode when using the unity editor. This change makes it so that during player builds (OnValidate is called during player builds probably to ensure that configurations are within their expected bounds), this isn't going to be invoked.

This does come down as a side effect of us using OnValidate (used to validate configs) for a purpose that it might not have originally been intended for (i.e. doing actual configuration work).

Also, updates the RegisterInstance function to find the first valid instance (during builds, objects will be destroyed without getting unregistered)