Update the MRTK to support being placed in the non-root Asset folder
https://github.com/microsoft/MixedRealityToolkit-Unity/issues/4864
https://github.com/microsoft/MixedRealityToolkit-Unity/issues/4067

Up until now, the MRTK has only worked if you had its code in the root Assets folder - there were specific parts of the code that would use Application.dataPath and do things relative to that, instead of using the MixedRealityToolkitFiles helper class to do correct folder resolution.

This fixes it so that the MRTK now SHOULD support other paths (i.e. if consumers want to place things in Assets/lib/mrtk/*), by updating the parts of the MRTK that currently use Asset absolute paths to instead use the MixedRealityToolkitFiles.

Also, adds tests to verify that MixedRealityToolkitFiles is actually able to reason over