Update the MRTK .json files to be pretty-formatted.
The current JSON files are really hard to review for deltas, because they are on such a massive long line that's it's impossible to verify any one specific change. This is especially difficult for reviewing in Github PR review tooling (or other ones that we're using)

This change does a few things:
1) Updates existing .json files to be pretty formatted.
2) Uploads the script that will pretty-print things.
3) Changes the locations in the code that generate these files to also pretty print.

Number 2 isn't actually necessary, but is still useful for future possible json stuff, and to have an easy way to correct things if the situation arises.