Update the build system to have a hook to check for validation errors in the MRTK build
Recently we hit an issue in the MRTK where we had duplicate GUIDs checked in for a few different assets:

https://github.com/microsoft/MixedRealityToolkit-Unity/pull/5431

The root issue here was that the meta files were copy pasted, instead of being regenerated at the new location. There was a warning in the console but it was ignored, and since Unity will non-deterministically choose one of them (i.e. either of the duplicate GUIDs could be right) - you end up with cases where you will get one or the other of the assets (and hopefully Unity choose correctly for you!)

I looked into doing a custom logger (i.e. catching LogWarning) but found that it doesn't run soon enough (i.e. warnings will show up, this code will get compiled) so this had to resort to editor-log parsing.

In theory this could be extended in the future to other issues that we hit that show up in the console as well.