Update the mrtk_pr pipeline to not build the UWP x86 build.
We currently build four different flavors on each mrtk_pr run, each of which can take between 12-17 minutes. This particular build being disabled in mrtk_pr has high correlation with the other x86 .NET build and IL2CPP ARM build.

Note that this is still enabled on the ongoing rolling build, which still needs to validate that everything is good.