Update the MRTK to support being placed in arbitrary folder names up to three directories deep.
In looking into https://github.com/microsoft/MixedRealityToolkit-Unity/issues/5568, I found that the existing MRTK files code would look for specific folder names that MRTK gets put into (as in, containing folder names). It had very specific behaviors with respect to NuGetForUnity or consumption as .unitypackage - if you had consumed the code some other way (for example, used a manual nuget restore step separately from NuGetForUnity) this wouldn't work out, because the naming of the folders would be different (i.e. MRTK/*) instead of (Microsoft.MixedReality.Toolkit*/*).

This change works around all that by introducing the concept of "sentinel" files, which are empty files that are named in a specific way and uniquely identify the root of an MRTK module/subfolder. Thus, when the MRTK files code finds a file called MRTK.SDK.sentinel, it knows that that folder is the root of the SDK MRTK subfolder, and it doesn't matter what name the root folder (or folders going up the root) are.