Add a code validation step to the PR pipeline.
We've seen a couple of instances of Boo.Lang getting checked in, where we really should be using System.Collections. More generally, there can be other patterns that we may check in inadvertently that we should avoid going forward, and having this be checked by a linter/during the checkin pipeline is better than assuming that all PR will catch it.