Clean up 'coming soon' documentation.
Some of these docs haven't been updated in a while, despite saying coming soon - as a principle, we should not check in things that are "coming soon" unless the work to do that has already been done and on your box locally and you're about to hit the button to create a PR. Otherwise, we risk creating a "graveyard" of documents that were dead before they even had the chance to live.

Some of these things being removed will be valuable to have docs on, and they can be re-added once they exist, but until then having empty pages has less value than nothing at all.