Remove PokeThreshold.
This variable was meaningless originally because it wasn't accounted for
in distance calculation. Fixing the distance calculation to use PokeThreshold
correctly was in turn leading to unintended large offsets and breaking all
kinds of established cases with buttons and other UX elements.

Since the touch system is already too complicated it was decided to remove
PokeThreshold altogether and restore the previous behavior.