Add a script that will check docs for validation errors.
https://github.com/microsoft/MixedRealityToolkit-Unity/issues/5716

Per the linked issue, there are currently a number of places in our docs that have hardcoded, full URLs to the github.io docs, instead of using relative paths. docfx is unable to validate full paths, but are able to validate relative paths, so in order to avoid inadvertently breaking things going forward (and to also ensure that things like doc versions properly point to their matching version), we want to make sure that no one checks in any more full paths.