New test for NearInteractionTouchables to make sure the poke distance… (#5264)
* New test for NearInteractionTouchables to make sure the poke distance behaves as expected.

* Fix outdated code in comments.

* Fix the poke pointer distance calculation to start touching at PokeThreshold.

* Override IsInteractionEnabled for the PokePointer to avoid raycasts outside TouchableDistance.

* Fix scale for UnityUI text boxes in tests.

* Fix for lost UnityUI mouse button changes when pointers are disabled.

This was because the MixedRealityInputModule was trying to dispatch one last
Unity mouse button update after pointers get disabled, but failed because
it was using the raycast result which was already null'd.

* Simplify TryRaisePokeDown/Up functions.

* Remove PokeThreshold.

This variable was meaningless originally because it wasn't accounted for
in distance calculation. Fixing the distance calculation to use PokeThreshold
correctly was in turn leading to unintended large offsets and breaking all
kinds of established cases with buttons and other UX elements.

Since the touch system is already too complicated it was decided to remove
PokeThreshold altogether and restore the previous behavior.