Turn off hand rays only when hand is near a grabbable object (#5823)
## Overview
If the hand is near the spatial mesh (SR), hand rays should not turn off. In fact, rays should only turn off if the hand is near a grabbable or touchable thing. This PR updates SpherePointer to return `IsNearObject` and `IsInteractionEnabled` only when the hand is near a collider that has a NearInteractionGrabbable on it. It also adds Grab Layer Masks, which allows developers to further prioritize and filter which objects a pointer can interact with. 

After this change, an object must both be on a Grabbable Layer as well as have a NearInteractionGrabbable component in order for hand rays to turn off. The Grabbable Layer is by default set to everything **except** Ignore Raycast and Spatial Awareness.

Fixes: #5748, #4203, #5621 

## Changes
- Add a Grabbable Layer Mask to Sphere Pointer to allow devs to easily disable all grabbable objects, and to improve performance
- The default Grabbale Layer Mask for the grab pointer is everything but Ignore Raycast and Spatial Awareness so that sphere casts don't check the spatial mesh (for perf).
- I also am caching the sphere query results to ensure we only do 3 sphere casts per frame (down from 11 without caching).
-  I snuck in a fix to pointer locking -- previously a pointer would lock when you clicked on any collidable, even if the collidable did not actually have any pointer handlers on it, or above it. now, the pointer will only lock if the pointer is actually grabbing an object that would handle pointer events. I may actually end up pulling that change into a separate PR.
- Add tests to verify my fix, and documentation reflecting the new grabbable behavior
- I also fixed the sphere pointer inspector -- it was actually not displaying several sphere pointer fields that it should have
- I removed the DebugSphere because it was not very useful, added a bit of extra code, and had bugs (did not turn off when debugsphere == false).

## Verification
- All playmode tests pass, new tests pass
- Verified that my new tests would have failed in old code
- Verified that my change actually fixes the bug (ray no longer hides when near SR)
- Runs on HoloLens 2 with no adverse perf impact.