Update the docs to address a term inconsistency in the MixedRealityConfigurationGuide
https://github.com/microsoft/MixedRealityToolkit-Unity/issues/5842

The usage of the word "pane" was confusing as highlighted in the issue - pane is technically correct here as well (it is, well, a pane) but looking through the rest of our docs, there isn't any other usage of the term pane so it is inconsistent. Switched to panel to be consistent with the rest of our docs.