Add build option for Unsafe Code (#5879)
This is a follow up that is useful for issue #5879. Research mode is usually used in combination with computer vision, for which performance is crucial. Therefor unsafe code for direct memory access is needed. With this changes unsafe code can simply be enabled with a checkbox in the unity build window. Like with research mode this checkbox is only visible if the build target is UWP and the targeted device is Hololens.