Add a CODEOWNERS to the MRTK.
CODEOWNERS is a handy way to both:
1. Provide insight into who knows most about an area of the code.
2. Automatically assign PRs to individuals based on previously determined ownership (i.e. the CODEOWNERS file)

This reduces the guesswork (i.e. "who should I include on this review?") and is especially interesting for external contributors (i.e. those who don't have rights to assign reviewers)

This change adds the beginnings of a CODEOWNERS file covering the relatively critical Build / CI area. This is an area that must be reviewed by someone who has been working on the Build / CI tooling.