Update the MRTK runtime to not allocate a new list on each Update loop.
https://github.com/microsoft/MixedRealityToolkit-Unity/issues/5942

The core MRTK update loop was allocating a new List<> (actually, two, if you look at both Update and LateUpdate). This is because each call to MixedRealityServiceRegistry wasn't being cached. In addition, LINQ adds its own GC overhead (i.e. after re-creating the same identical list, the same list was then sorted again... and again... and again...)

This change avoids that by updating the GetAllServices call to return a cached list of services, which are sorted in priority order. That way the MRTK instead of having to re-copy and resort, can just use the same list on every single frame.

I had also considered keeping the original "sort order" (i.e. random, or actually, insertion order into the registry) but we never formally defined what the ordering of elements was here, and the prime and most important scenario here depends on it being priority order sorted, so we may as well optimize for that use case.

I didn't fully get rid of LINQ (note that the Reverse operation still uses LINQ) - this is okay because it's only in the shutdown path. Alternatively I could have changed the function signature of GetAllServices to be IReadOnlyList, but this would have been a breaking change for not a lot of value.