Update the AsyncCoroutineRunner to be properly rooted for DontDestroyOnLoad
https://github.com/microsoft/MixedRealityToolkit-Unity/issues/5900

DontDestroyOnLoad is only valid to be called on an object that is at the root (i.e. it has a null parent). It's possible for the AsyncCoroutineRunner to have been attached to the mixed reality playspace (i.e. if the user manually created one under the playspace or under any other object).

This updates the code to do a one-time at runtime fixup of its location to make sure that it can properly be DontDestroyOnLoad-ed.