Fix the HandConstraint solver to only consider controllers with handedness != None to be valid targets.
https://github.com/microsoft/MixedRealityToolkit-Unity/issues/5749

The Xbox controller is an IMixedRealityController, that, when plugged in, will trigger the hand menu because it's considered a valid target for the hand constraint solver.

This change makes it so that the solver will filter out any controllers that don't have a handedness specified (i.e. they should be left/right/any/both)