Add translation constraints to the Manipulation Handler
https://github.com/microsoft/MixedRealityToolkit-Unity/issues/4154

This change introduces movement constraints on the manipulation handler, so that objects can be configured to only move in a specific axis or plane.

This is implemented with a "permissive" model (i.e. an object can be configured to have free movement on certain axis). The set of free axes are configured via a flag-based parameter (Translation Axes).

Adds some test coverage (not fully exhaustive, just checking the main cases i.e. each axis and a single plane)

Other thoughts:
It's kinda weird that you can select X, Y, and Z at the same time (basically no-oping the constraints). I wanted to add a more specific warning here but without this PR in (https://github.com/microsoft/MixedRealityToolkit-Unity/pull/5955/files) it would involve adding a custom inspector as well. First it's not that big a deal, but also if we want to deal with it once the other PR is in that would be fairly easy (just add a warning block to the inspector code + conditional showing of the parameter)

I opted to not bake this into ConstraintOnMovement because I didn't want to explode a bunch of different options (i.e. X only, Y only, Z only, XY only, YZ only, XZ only), and I think that this general type of "Axis Constraint" exists elsewhere (it does if we look at rotation). I opted to name this somewhat generically so that it can be used in other scenarios that have axis-flag type behaviors.