Turn off input simulation indicators (#6062)
## Overview
UPDATE: This PR originally moved input simulation indicators to the sides and updates the reset button to look like a 'refresh' button. However, after moving indicators to the side, raycasts related to Unity UI began failing. Since the indicators as they are are too visually distracting from interaction, check them in as off by default. Filed #6106 to track turning them back on. The indicators as they are checked in will be on the sides, with correct refresh buttons.

![image](https://user-images.githubusercontent.com/168492/65546816-198f6280-decd-11e9-8b7f-39b123f19dc4.png)


## Changes
- Fixes: #6023