removed patch for scaling issue while animating because it's breaking scaling of bounding box children (#6119)
fix for #6088

I removed a fix I did to compensate for a messed up scaling caused by animating the localscale of the bounding box target object. ( #6017 )
This fix broke scaling behavior for bounding box children.
Seems like scale changes to children are not propagated anymore once I set the localScale manually. This didn't show in the objects I tested with because the scaled objects were the target object itself. The broken cheese however has the scaled mesh as a child (same goes for slate, coffee, etc).

I tried to add a different workaround for the messed up scaling (eg https://answers.unity.com/questions/770240/changing-the-scale-of-the-parent-messes-up-the-chi.html) but that didn't work with the test file I'm using from one of the partners.
While testing with my own animation files this behavior won't show.

I don't think the original issue #5006 is something we can fix in MRTK code (the exploding part, not the updating part -> that's already fixed).

nit: added missing parentheses to if