Auto generate signing configs in release pipeline
The signing configs will go out of sync any time add, remove or rename assemblies. We previously created a script to generate these configs, as writing them manually was a pain. It makes sense that we use this script in the pipelines as it's one less thing to think about.