Rigidbody colliding with own BoundingBox handles
In cases where BoundingBox uses a child object as its target, and the child was a Rigidbody, the child object would collide with the BoundingBox handles, causing the object to jump about.
This change adds the BoundingBox handles as children of the target object, rather than the BoundingBox gameObject. This makes the target collider and the handler colliders into a compound collider that does not collide with itself.