Add build tooling to ensure that we don't check in more hardcoded paths
We've fixed a number of hardcoded path related issues in the past, and we're still seeing them get checked in. This adds a verifier to ensure that we don't add new ones, while still having a way to have exceptions in cases where they are actually legitimate.

This has exceptions for the existing custom profile generator and test code, along with the scene system .unity files that are breaking arbitrary folder moves.