Fix spelling errors and dead links in the coding guideslines.
I was reading this as part of reviewing another change, and noticed some immediate spelling errors really early on in the doc, which can reduce confidence on the quality of the rest of it (i.e. if the first thing you see is a spelling mistake in a long document, it can make you go like... huh...)

There was also a broken link (to CI) that was referencing a specific build from ages ago - that build information has long been cleared, so it was a dead link. I've updated that link to point the mrtk_pr build (which I believe is the most relevant given the other context in that area)