Fix SR_Triangles and SR_Triangles_Object_Space shader warnings
https://github.com/microsoft/MixedRealityToolkit-Unity/issues/6609

The original change that added these shaders had a few warnings:
https://github.com/keveleigh/HoloToolkit-Unity/commit/6fa80b156b0c6138efd64d82c585f1e742812016#diff-424c2d8798826167e9e8bbc3c1a6dd70

In particular, there was some code that added while loop that would always end up running once (i.e. it would be equivalent to just not having the loop). This gives a warning in Unity, and the MRTK should be warning-free. This work just unrolls the loop manually (i.e. recognizing that it is only ever run once, just getting rid of the loop)