Fix possible follow object stuck behind player
Discovered a possible issue with the follow solver. Although it was hard to repro, it was possible to get the follow object stuck in the inverted leashing frustum. This fix uses the angle from `currentRefForward` and `toTarget` projected onto the reference yz plane for vertical leashing, but use the angle from `toTarget` to the reference yz plane for horizontal leashing. This ensures that positions in the inverted leashing frustum are not valid leasing positions.