Initial code to fix Issue 8344: Add reading mode support to the Unity camera for HoloLens 2
This code adds a "projection override" mode to the main camera.  It reduces the system's FOV but eliminates a scaling of Unity's output.  So a pixel rendered by Unity will correspond to a projected pixel on HoloLens 2.  Application authors should do tests with multiple individuals to be sure this is a tradeoff they want in their app.

This mode is not something you want enabled for the lifetime of the app.  It should be enabled when necessary - only when smaller details need to be sharper.  This mode should stay disabled otherwise to maintain then larger FOV

There is currently a bug (as of Unity 2019.4.5) where Unity doesn't fully use the adjusted projection when the "Stereo Rendering Mode" is "Single Pass Instanced."  Until that is fixed, you can work around it by changing to "Multi Pass."  This mode will have some performance impact for the entire app experience, not just when Projection Override is enabled.

Code changes:
* Add ProjectionOverrideEnabled to IMixedRealityCameraSystem and implement it in MixedRealityCameraSystem.
* Add ProjectionOverride MonoBehavior that does the actual Projection Override
* Add ProjectionOptionsDemo scene to demonstrate the change.