Fix Shader import error seen on Mac editor
OnLoadUtilities.cs fails to copy shaders when running on MacOS due to
a filename sensitivity issue.

The constant `ShaderSentinelFile = "MRTK.Shaders.Sentinel";` refers to
a file that is actually named `MRTK.Shaders.sentinel` with a lowercase
`s` on sentinel.

On MacOS DirectoryInfo.GetFiles appears to be case-sensitive even if
the underlying filesystem is not. This causes the script to not find
MRTK.Shaders.sentinel and thus not be able to copy the shaders.

Fix the issue by using a lowercase `s` in the constant.

Fixes #8826