Fix hand meshes popping in and out quickly as hand tracking is regained (#9736)
RiggedHandVisualizer is instantiated when hand tracking is regained. At
that point, the HandRenderer is enabled and actually rendering, making
the hand mesh visible - even if hand mesh visualization is disabled in
MRTK scene settings.

Afterwards, OnHandJointsUpdated is called in a subsequent frame, which
will evaluate if the profile has hand mesh visualisation enabled, and
disable it if not. This results in a bug where, if you had hand mesh
visualization enabled, you see the hand mesh popping in at its original
location, and then see it disappear again because it is being hidden. If
you had hand mesh visualization enabled, you would also see this
popping, because the location of the mesh wasn't updated until later.

Disabling the hand mesh rendering on start (until we've been able to
update its position, and evaluate if it needs to render at all) seems
to fix this.