Add option to move grasp point to fingertip
The grasp pointer by default is positioned between the thumb and index finger. However this is confusing if a fingertip cursor is rendered (either because one is added to the grasp pointer or because grabbable elements are near pokable ones). This adds an option to move the grasp point to the fingertip if desired.