Fixed log spam that is coming from HandConstraintPalmUp.cs
The HandConstraintPalmUp.cs script is logging a warning whenever it cannot find the palm joint. However this scenario happens quite often when the HL detects a partial hand with palm joint missing. In that scenario, the hand input is considered invalid which causes no harm at all.

The solution is to remove the debug log for this legitimate and common scenario to not flood our logs with unnecessary warnings.