Fix for flickering Proximity Lighting
We notice a small but distracting flicker when moving from a poke to a grab pointer when pokable and grabbable objects are in close proximity. There's a single frame when there are no lights enabled. Adding a forced update to `OnEnable()` (similarly to the behaviour of `OnDisable()`) fixes the issue.