Prevent PokePointer from returning zero-vector normals
The implementation of `TryGetNormalToNearestSurface()` for `PokePointer` always returned true, even when no proximate collider was found. This could cause zero-vectors to be returned by that function, which created mathematical errors and log spam down the line.

I've altered this implementation to return false when there's no nearby collider and set the output to a valid unit vector regardless (this is the same output as `GrabPointer` in this situation).